requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
aiohttp>=3.8.0
tqdm>=4.65.0
tabulate>=0.9.0
pyyaml>=6.0
